Task : Take the Model class file and break down into separate files for model, entity, context, and add interfaces to public classes.
