+++ Context/CitiesContext.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Data.Entity;
+using Entity;
+
+namespace Context
+{
+    public class CitiesContext : DbContext
+    {
+        public DbSet<CountryEntity> Countries { get; set; }
+        public DbSet<CityEntity> Cities { get; set; }
+    }
+}
