+++ Entity/CityEntity.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.ComponentModel.DataAnnotations;
+using Model;
+
+namespace Entity
+{
+    public class CityEntity : CityModel
+    {
+        [Key]
+        public int CityID { get; set; }
+
+        public virtual CountryEntity Country { get; set; }
+    }
+}
