+++ Entity/CountryEntityQueries.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Entity
+{
+    public static class CountryEntityQueries
+    {
+
+        
+    }
+}
