+++ Model/CityModel.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace Model
+{
+    public class CityModel
+    {
+            public string City_name { get; set; }
+            public string City_ascii { get; set; }
+            public double Lat { get; set; }
+            public double Lng { get; set; }
+            public string Admin_name { get; set; }
+            public string Capital { get; set; }
+            public double Population { get; set; }
+            public int CountryId { get; set; }
+        
+    }
+}
