+++ Model/CountryModel.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Model
+{
+    public class CountryModel
+    {
+        public string Name { get; set; }
+        public string ISO2 { get; set; }
+        public string ISO3 { get; set; }
+    }
+}
